+++ lib/Event/ConfigureOptionsEvent.php
@@ -12,10 +12,8 @@
 namespace Sulu\Component\DocumentManager\Event;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Sulu\Component\DocumentManager\DocumentManagerContext;
-use Symfony\Component\EventDispatcher\Event;
 
-class ConfigureOptionsEvent extends Event
+class ConfigureOptionsEvent extends AbstractEvent
 {
     use EventOptionsTrait;
