+++ lib/Event/QueryCreateBuilderEvent.php
@@ -14,7 +14,7 @@ namespace Sulu\Component\DocumentManager\Event;
 use Sulu\Component\DocumentManager\Exception\DocumentManagerException;
 use Sulu\Component\DocumentManager\Query\QueryBuilder;
 
-class QueryCreateBuilderEvent extends AbstractEvent
+class QueryCreateBuilderEvent extends AbstractDocumentManagerContextEvent
 {
     /**
      * @var
