+++ lib/Event/QueryExecuteEvent.php
@@ -12,10 +12,10 @@
 namespace Sulu\Component\DocumentManager\Event;
 
 use Sulu\Component\DocumentManager\Collection\QueryResultCollection;
-use Sulu\Component\DocumentManager\Query\Query;
 use Sulu\Component\DocumentManager\DocumentManagerContext;
+use Sulu\Component\DocumentManager\Query\Query;
 
-class QueryExecuteEvent extends AbstractEvent
+class QueryExecuteEvent extends AbstractDocumentManagerContextEvent
 {
     use EventOptionsTrait;
 
@@ -30,6 +30,7 @@ class QueryExecuteEvent extends AbstractEvent
     private $result;
 
     /**
+     * @param DocumentManagerContext $context
      * @param Query $query
      * @param array $options
      */
