+++ tests/Bootstrap.php
@@ -13,8 +13,6 @@ namespace Sulu\Component\DocumentManager\tests;
 
 use Jackalope\RepositoryFactoryDoctrineDBAL;
 use Jackalope\Transport\DoctrineDBAL\RepositorySchema;
-use Monolog\Handler\StreamHandler;
-use Monolog\Logger;
 use PHPCR\SessionInterface;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -25,16 +23,8 @@ class Bootstrap
 {
     public static function createContainer()
     {
-        $logDir = __DIR__ . '/../data/logs';
-
-        if (!file_exists($logDir)) {
-            mkdir($logDir);
-        }
-
         $container = new ContainerBuilder();
         $container->set('doctrine_phpcr.default_session', self::createSession());
-        $logger = new Logger('test');
-        $logger->pushHandler(new StreamHandler($logDir . '/test.log'));
 
         $dispatcher = new ContainerAwareEventDispatcher($container);
         $container->set('sulu_document_manager.event_dispatcher', $dispatcher);
@@ -145,7 +135,6 @@ class Bootstrap
             'host' => 'localhost',
             'user' => 'admin',
             'password' => 'admin',
-            //'path' => __DIR__ . '/../data/test.sqlite',
             'memory' => true,
         ]);
